abstracting the build scripts to not be ionic framework exclusive
