+++ scripts/npm.ts
@@ -5,15 +5,15 @@ import * as fs from 'fs';
 import * as config from './config';
 import * as utils from './utils';
 
-export async function install() {
+export async function installAPI() {
   utils.vlog('npm installing monorepo');
-  await utils.shell('cd ionic && npm ci');
+  await utils.shell(`cd ${config.IONIC_DIR} && npm ci`);
   utils.vlog('npm installing core');
-  return await utils.shell('cd ionic/core && npm ci');
+  return await utils.shell(`cd ${config.IONIC_DIR}/core && npm ci`);
 }
 
 export async function buildAPIDocs() {
   utils.vlog('building');
-  await utils.shell('cd ionic/core && npm run build.docs.json');
-  return JSON.parse(fs.readFileSync('ionic/core/dist/docs.json', 'utf8'));
+  await utils.shell(`cd ${config.IONIC_DIR}/core && npm run build.docs.json`);
+  return JSON.parse(fs.readFileSync(`${config.IONIC_DIR}/core/dist/docs.json`, `utf8`));
 }
